update TmSessionWelcome tests
